Move `EntityStore` to `store`
