+++ apps/hash-graph/src/subcommand/reindex_cache.rs
@@ -1,11 +1,10 @@
 use clap::Parser;
 use error_stack::{Report, ResultExt as _, ensure};
-use graph::store::{
-    DatabaseConnectionInfo, DatabasePoolConfig, PostgresStorePool, StorePool as _,
-    knowledge::EntityStore as _,
-};
+use graph::store::{DatabaseConnectionInfo, DatabasePoolConfig, PostgresStorePool, StorePool as _};
 use hash_graph_authorization::NoAuthorization;
-use hash_graph_store::{data_type::DataTypeStore as _, entity_type::EntityTypeStore as _};
+use hash_graph_store::{
+    data_type::DataTypeStore as _, entity::EntityStore as _, entity_type::EntityTypeStore as _,
+};
 use tokio_postgres::NoTls;
 
 use crate::error::GraphError;
