+++ libs/@local/graph/api/src/rest/entity.rs
@@ -12,15 +12,8 @@ use axum::{
 };
 use error_stack::{Report, ResultExt as _};
 use graph::store::{
-    NullOrdering, Ordering, StorePool,
+    StorePool,
     error::{EntityDoesNotExist, RaceConditionOnUpdate},
-    knowledge::{
-        ClosedMultiEntityTypeMap, CountEntitiesParams, CreateEntityRequest, DiffEntityParams,
-        DiffEntityResult, EntityQueryCursor, EntityQuerySorting, EntityQuerySortingRecord,
-        EntityStore as _, EntityValidationType, GetEntitiesParams, GetEntitiesResponse,
-        GetEntitySubgraphParams, PatchEntityParams, QueryConversion, UpdateEntityEmbeddingsParams,
-        ValidateEntityParams,
-    },
 };
 use hash_graph_authorization::{
     AuthorizationApi as _, AuthorizationApiPool,
@@ -34,9 +27,17 @@ use hash_graph_authorization::{
 };
 use hash_graph_store::{
     account::AccountStore as _,
-    entity::{EntityQueryPath, EntityQuerySortingToken, EntityQueryToken},
+    entity::{
+        ClosedMultiEntityTypeMap, CountEntitiesParams, CreateEntityRequest, DiffEntityParams,
+        DiffEntityResult, EntityQueryCursor, EntityQueryPath, EntityQuerySorting,
+        EntityQuerySortingRecord, EntityQuerySortingToken, EntityQueryToken, EntityStore as _,
+        EntityValidationType, GetEntitiesParams, GetEntitiesResponse, GetEntitySubgraphParams,
+        PatchEntityParams, QueryConversion, UpdateEntityEmbeddingsParams, ValidateEntityComponents,
+        ValidateEntityParams,
+    },
     entity_type::{EntityTypeResolveDefinitions, IncludeEntityTypeOption},
     filter::Filter,
+    query::{NullOrdering, Ordering},
     subgraph::{edges::GraphResolveDepths, temporal_axes::QueryTemporalAxesUnresolved},
 };
 use hash_graph_types::{
@@ -61,7 +62,6 @@ use hash_graph_types::{
     },
     owned_by_id::OwnedById,
 };
-use hash_graph_validation::ValidateEntityComponents;
 use hash_temporal_client::TemporalClient;
 use serde::{Deserialize, Serialize};
 use type_system::url::VersionedUrl;
