+++ libs/@local/graph/graph/src/store/postgres/ontology/data_type.rs
@@ -19,6 +19,7 @@ use hash_graph_store::{
     },
     error::{InsertionError, QueryError, UpdateError},
     filter::{Filter, FilterExpression, ParameterList},
+    query::VersionedUrlSorting,
     subgraph::{
         Subgraph, SubgraphRecord as _,
         edges::{EdgeDirection, GraphResolveDepths, OntologyEdgeKind},
@@ -53,7 +54,7 @@ use type_system::{
 };
 
 use crate::store::{
-    crud::{QueryResult as _, Read, ReadPaginated, VersionedUrlSorting},
+    crud::{QueryResult as _, Read, ReadPaginated},
     error::DeletionError,
     postgres::{
         AsClient, PostgresStore, TraversalContext,
