+++ libs/@local/graph/store/src/entity/mod.rs
@@ -1,6 +1,19 @@
-pub use self::query::{EntityQueryPath, EntityQuerySortingToken, EntityQueryToken};
+pub use self::{
+    query::{
+        EntityQueryCursor, EntityQueryPath, EntityQuerySorting, EntityQuerySortingRecord,
+        EntityQuerySortingToken, EntityQueryToken,
+    },
+    store::{
+        ClosedMultiEntityTypeMap, CountEntitiesParams, CreateEntityParams, CreateEntityRequest,
+        DiffEntityParams, DiffEntityResult, EntityStore, EntityValidationType, GetEntitiesParams,
+        GetEntitiesResponse, GetEntitySubgraphParams, GetEntitySubgraphResponse, PatchEntityParams,
+        QueryConversion, UpdateEntityEmbeddingsParams, ValidateEntityComponents,
+        ValidateEntityError, ValidateEntityParams,
+    },
+};
 
 mod query;
+mod store;
 
 use hash_graph_types::knowledge::entity::Entity;
