+++ libs/@local/graph/store/src/entity_type/store.rs
@@ -24,9 +24,9 @@ use type_system::{
 };
 
 use crate::{
-    ConflictBehavior,
     error::{InsertionError, QueryError, UpdateError},
     filter::Filter,
+    query::ConflictBehavior,
     subgraph::{Subgraph, edges::GraphResolveDepths, temporal_axes::QueryTemporalAxesUnresolved},
 };
