+++ libs/@local/graph/store/src/lib.rs
@@ -6,8 +6,6 @@
 
 extern crate alloc;
 
-use serde::{Deserialize, Serialize};
-
 pub mod account;
 pub mod data_type;
 pub mod entity;
@@ -17,12 +15,5 @@ pub mod property_type;
 pub mod error;
 pub mod filter;
 pub mod migration;
+pub mod query;
 pub mod subgraph;
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
-pub enum ConflictBehavior {
-    /// If a conflict is detected, the operation will fail.
-    Fail,
-    /// If a conflict is detected, the operation will be skipped.
-    Skip,
-}
