+++ tests/graph/benches/representative_read/seed.rs
@@ -1,12 +1,12 @@
 use core::{iter::repeat, str::FromStr as _};
 use std::collections::{HashMap, HashSet, hash_map::Entry};
 
-use graph::store::{
-    AsClient as _,
-    knowledge::{CreateEntityParams, EntityStore as _},
-};
+use graph::store::AsClient as _;
 use hash_graph_authorization::{AuthorizationApi, schema::WebOwnerSubject};
-use hash_graph_store::account::{AccountStore as _, InsertAccountIdParams, InsertWebIdParams};
+use hash_graph_store::{
+    account::{AccountStore as _, InsertAccountIdParams, InsertWebIdParams},
+    entity::{CreateEntityParams, EntityStore as _},
+};
 use hash_graph_test_data::{data_type, entity, entity_type, property_type};
 use hash_graph_types::{
     account::AccountId,
