+++ tests/graph/integration/postgres/drafts.rs
@@ -1,7 +1,7 @@
 use std::collections::HashSet;
 
-use graph::store::knowledge::{CreateEntityParams, EntityStore as _, PatchEntityParams};
 use hash_graph_authorization::AuthorizationApi;
+use hash_graph_store::entity::{CreateEntityParams, EntityStore as _, PatchEntityParams};
 use hash_graph_temporal_versioning::ClosedTemporalBound;
 use hash_graph_test_data::{data_type, entity, entity_type, property_type};
 use hash_graph_types::{
