+++ tests/graph/integration/postgres/property_metadata.rs
@@ -2,8 +2,8 @@ use alloc::borrow::Cow;
 use core::{iter::once, str::FromStr as _};
 use std::collections::{HashMap, HashSet};
 
-use graph::store::knowledge::{CreateEntityParams, EntityStore as _, PatchEntityParams};
 use hash_graph_authorization::AuthorizationApi;
+use hash_graph_store::entity::{CreateEntityParams, EntityStore as _, PatchEntityParams};
 use hash_graph_test_data::{data_type, entity, entity_type, property_type};
 use hash_graph_types::{
     knowledge::{
